Coding, Emscripten - fix runtime crashes on WASM64 (-sMEMORY64=1)
by explicitly casting between BigInt and Number within EM_JS() blocks.

+++ src/OSD/OSD_MemInfo.cxx
@@ -34,7 +34,7 @@
   #include <emscripten.h>
 
   //! Return WebAssembly heap size in bytes.
-  EM_JS(size_t, OSD_MemInfo_getModuleHeapLength, (), {
+  EM_JS(double, OSD_MemInfo_getModuleHeapLength, (), {
     return Module.HEAP8.length;
   });
 #endif
@@ -170,7 +170,7 @@ void OSD_MemInfo::Update()
   }
   if (IsActive (MemVirtual))
   {
-    myCounters[MemVirtual] = OSD_MemInfo_getModuleHeapLength();
+    myCounters[MemVirtual] = (size_t)OSD_MemInfo_getModuleHeapLength();
   }
 #elif (defined(__linux__) || defined(__linux))
   if (IsActive (MemHeapUsage))
